refactoring cryptoUtil by removing all common methods. All the removed methods are added to CommonUtils.js for standardization
